change brush

+++ src/paintfield/extensions/brushtool/brushstrokersimplebrush.h
@@ -23,7 +23,7 @@ protected:
 private:
 	
 	double drawSegment(const Malachite::Vec2D &p1, const Malachite::Vec2D &p2, double len, double &pressure, double pressureNormalized, double carryOver);
-	QRect drawDab(const Malachite::Vec2D &pos, double pressure);
+	QRect drawDab(const Malachite::Vec2D &pos, float pressure);
 	
 	void drawScanline(const BrushScanline &scanline, Malachite::Surface *surface);
